langandsubtitle/app/requirements.txt
langdetect
vosk
moviepy
speechRecognition
python-engineio==3.13.2
python-socketio==4.6.1
